pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portals-gifts-bot"
version = "2.0.0"
description = "Telegram bot for monitoring gifts on marketplaces"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.11.0",
    "aiomysql>=0.2.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "aportalsmp>=1.3.0",
    "pyrogram>=2.0.0",
    "tonnelmp>=1.0.0",
]

[project.optional-dependencies]
redis = ["redis[hiredis]>=5.0.0"]
webhook = ["aiohttp>=3.9.0"]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]


